Move handleRequestFailure to request.ts - prevent circular dependency
