+++ packages/components/src/internal/actions.ts
@@ -39,7 +39,7 @@ import {
 import { DataViewInfo } from './DataViewInfo';
 import { getQueryParams } from './util/URL';
 
-import { handleRequestFailure } from './util/utils';
+import { handleRequestFailure } from './request';
 import { resolveErrorMessage } from './util/messaging';
 import { buildURL } from './url/AppURL';
