+++ packages/components/src/internal/components/security/APIWrapper.ts
@@ -13,7 +13,8 @@ import { Principal, SecurityPolicy, SecurityRole } from '../permissions/models';
 import { selectRows } from '../../query/selectRows';
 import { SCHEMAS } from '../../schemas';
 import { naturalSortByProperty } from '../../../public/sort';
-import { caseInsensitive, handleRequestFailure } from '../../util/utils';
+import { caseInsensitive } from '../../util/utils';
+import { handleRequestFailure } from '../../request';
 import { getUserProperties } from '../user/actions';
 import { flattenValuesFromRow } from '../../../public/QueryModel/QueryModel';
 import { SchemaQuery } from '../../../public/SchemaQuery';
