+++ packages/components/src/internal/request.ts
@@ -1,6 +1,36 @@
 import { Ajax, Utils } from '@labkey/api';
 
-import { handleRequestFailure } from './util/utils';
+/**
+ * Handles Ajax.request() failures. Calls the passed in "reject" handler with the error
+ * from the request as determined by Utils.getCallbackWrapper(). The "reject" parameter can be a normal function
+ * or a rejection handler for a Promise. If a response status code is available, then it will append that value
+ * to the error object as "status".
+ * Example:
+ * ```
+ * import { Ajax } from '@labkey/api';
+ *
+ * new Promise((resolve, reject) => {
+ *     return Ajax.request({
+ *         // ... url, success handler, etc
+ *         failure: handleRequestFailure(reject, 'This optional message is logged to console.error'),
+ *     });
+ * });
+ * ```
+ */
+export function handleRequestFailure(reject: (error: any) => void, logMsg?: string) {
+    return Utils.getCallbackWrapper(
+        (error, response) => {
+            // Appends the response's status code to the error object
+            const errorWithStatus = { ...error, status: response?.status };
+            if (logMsg) {
+                console.error(logMsg, errorWithStatus);
+            }
+            reject(errorWithStatus);
+        },
+        undefined,
+        true
+    );
+}
 
 export type RequestHandler = (request: XMLHttpRequest) => void;
